\name{fit_GLMs}
\alias{fit_GLMs}
\title{
Fit GLM models for imputed or synthetic datasets
}
\description{
Fit GLM models for imputed or synthetic datasets
}
\usage{
fit_GLMs(InputData, exp)
}
\arguments{
  \item{InputData}{a list of imputed or synthetic datasets}
  \item{exp}{GLM expression (for polr and nnet, those libraries should be loaded first)}
}
\value{
Results: a list of GLM results
}
