\name{MCZ}
\alias{MCZ}
\title{
Example dataframe for structrual zeros based on the NYMockexample dataset.
}
\description{
Example dataframe for structrual zeros based on the NYMockexample dataset. It contains 8 structural zero cases with 10 variables.
  \tabular{rl}{
    [,1] \tab AGE = 15 and EDUC = 8\cr
    [,2] \tab AGE = 16 and VESTAT = 2\cr
    [,3] \tab OWNERSHIP = 0 and MORTGAGE = 4\cr
    [,4] \tab AGE = 17 and EDUC = 11 \cr
    [,5] \tab AGE = [36, 50] and EMPSTAT = 0\cr
    [,6] \tab AGE > 70 and DISABWRK = 0\cr
    [,7] \tab AGE < 15 and EDUC = 10 \cr
    [,8] \tab OWNERSHIP = 2 and MORTGAGE = 1 \cr
  }
}
