\name{ss16pusa_mi_MCZ}
\alias{ss16pusa_mi_MCZ}
\title{
Example dataframe for structrual zeros based on the ss16pusa_sample_zeros dataset.
}
\description{
Example dataframe for structrual zeros based on the ss16pusa_sample_zeros dataset. It contains 8 structural zero cases with 5 variables.
\tabular{rl}{
    [,1] \tab AGEP = 16 and SCHL = Bachelor's degree\cr
    [,2] \tab AGEP = 16 and SCHL = Doctorate degree\cr
    [,3] \tab AGEP = 16 and SCHL = Master's degree\cr
    [,4] \tab AGEP = 16 and SCHL = Professional degree \cr
    [,5] \tab AGEP = 17 and SCHL = Bachelor's degree\cr
    [,6] \tab AGEP = 17 and SCHL = Doctorate degree\cr
    [,7] \tab AGEP = 17 and SCHL = Master's degree \cr
    [,8] \tab AGEP = 17 and SCHL = Professional degree \cr
}
}
