\name{ss16pusa_sample_nozeros}
\alias{ss16pusa_sample_nozeros}

\title{
Example dataframe for input categorical data without structural zeros (without missing values).
}
\description{
Example dataframe for input categorical data without structural zeros (without missing values). It contains 1000 observations and 3 variables.
\tabular{rlll}{
    [,1] \tab MAR \tab marital status \tab 5 levels \cr
    [,2] \tab SEX    \tab sex \tab 2 levels \cr
    [,3] \tab WKL   \tab When last worked \tab 3 levels \cr
}
}
