\name{ss16pusa_sample_nozeros_miss}
\alias{ss16pusa_sample_nozeros_miss}

\title{
Example dataframe for input categorical data without structural zeros (with missing values).
}
\description{
Example dataframe for input categorical data without structural zeros (with missing values). It contains 1000 observations and 3 variables.
\tabular{rlll}{
    [,1] \tab MAR \tab marital status \tab 5 levels: Married; Widowed; Divorced; Separated; Never married. \cr
    [,2] \tab SEX    \tab sex \tab 2 levels: Male; Female. \cr
    [,3] \tab WKL   \tab When last worked \tab 3 levels: Within the last 12 months; 1-5 years ago; \cr
         \tab       \tab                 \tab  Over 5 years ago or never worked. \cr
}
}
