\name{ss16pusa_sample_zeros}
\alias{ss16pusa_sample_zeros}

\title{
Example dataframe for input categorical data with structural zeros (without missing values).
}
\description{
Example dataframe for input categorical data with structural zeros (without missing values). It contains 1000 observations and 5 variables.
\tabular{rlll}{
    [,1] \tab AGEP  \tab age \tab 7 levels \cr
    [,2] \tab MAR \tab marital status \tab 5 levels \cr
    [,3] \tab SCHL    \tab educational attainment \tab 9 levels \cr
    [,4] \tab SEX    \tab sex \tab 2 levels \cr
    [,5] \tab WKL   \tab When last worked \tab 3 levels \cr
}
}
