\name{X}
\alias{X}

\title{
Example dataframe for input categorical data with missing values based on the NYMockexample dataset.
}
\description{
Example dataframe for input categorical data with missing values based on the NYMockexample dataset. It contains 2000 observations and 10 variables.
\tabular{rlll}{
    [,1] \tab OWNERSHIP  \tab ownership of dwelling \tab 3 levels \cr
    [,2] \tab MORTGAGE \tab mortgate status \tab 4 levels \cr
    [,3] \tab AGE    \tab age \tab 9 levels \cr
    [,4] \tab SEX    \tab sex \tab 2 levels \cr
    [,5] \tab MARST   \tab martial status \tab 6 levels \cr
    [,6] \tab RACESING     \tab single race identification \tab 5 levels \cr
    [,7] \tab EDUC        \tab educational attainment \tab 11 levels \cr
    [,8] \tab EMPSTAT   \tab employment status \tab 4 levels \cr
    [,9] \tab DISABWRK \tab work disability status \tab 3 levels \cr
    [,10] \tab VESTAT \tab veteran status \tab 3 levels \cr
}
}

