\name{ss16pusa_sample_zeros}
\alias{ss16pusa_sample_zeros}

\title{
Example dataframe for input categorical data with structural zeros (without missing values).
}
\description{
Example dataframe for input categorical data with structural zeros (without missing values). It contains 1000 observations and 5 variables.
\tabular{rlll}{
    [,1] \tab AGEP  \tab age \tab 7 levels: 16; 17; [18, 24]; [25, 35]; [36, 50]; [51, 70]; (70, ). \cr
    [,2] \tab MAR \tab marital status \tab 5 levels: Married; Widowed; Divorced; Separated; \cr
     \tab \tab  \tab Never married. \cr
    [,3] \tab SCHL    \tab educational attainment \tab 9 levels: Up to K0; Some K12, no diploma;   \cr
     \tab  \tab  \tab  High school diploma or GED; Some college, no degree;   \cr
      \tab  \tab  \tab  Associate's degree; Bachelor's degree; Master's degree;  \cr
       \tab  \tab  \tab Professional degree; Doctorate degree. \cr
    [,4] \tab SEX    \tab sex \tab 2 levels: Male; Female. \cr
    [,5] \tab WKL   \tab When last worked \tab 3 levels: Within the last 12 months; 1-5 years ago;  \cr
     \tab  \tab  \tab  Over 5 years ago or never worked. \cr
}
}
