\name{X}
\alias{X}

\title{
Example dataframe for input categorical data with missing values based on the NYMockexample dataset.
}
\description{
Example dataframe for input categorical data with missing values based on the NYMockexample dataset. It contains 2000 observations and 10 variables.
\tabular{rlll}{
    [,1] \tab OWNERSHIP  \tab ownership of dwelling \tab 3 levels: N/A; Owned or being bought (loan); \cr
     \tab   \tab  \tab  Rented. \cr
    [,2] \tab MORTGAGE \tab mortgate status \tab 4 levels: N/A; No, owned free and clear; \cr
     \tab  \tab  \tab  Yes, mortgaged / deed of trust or similar debt;  \cr
     \tab  \tab  \tab  Yes, contract to purchase. \cr
    [,3] \tab AGE    \tab age \tab 9 levels: [0, 14]; 15; 16; 17; [18, 24]; [25, 35]; [36, 50];  \cr
     \tab    \tab \tab 9  [51, 70]; [71, ). \cr
    [,4] \tab SEX    \tab sex \tab 2 levels: Male; Female. \cr
    [,5] \tab MARST   \tab martial status \tab 6 levels: Married, spouse present; Married, spouse absent; \cr
     \tab   \tab  \tab Separated; Divorced; Widowed; Never married / single. \cr
    [,6] \tab RACESING     \tab single race identification \tab 5 levels: White; Black; American Indian / Alaska Native;\cr
     \tab     \tab  \tab   Asian and / or Pacific Islander; Other race, non-Hispanic. \cr
    [,7] \tab EDUC        \tab educational attainment \tab 11 levels: N/A or no schooling; Nursery school to grade 4;  \cr
      \tab     \tab   \tab  Grade 5, 6, 7, or 8; Grade 9; Grade 10; Grade 11; \cr
      \tab     \tab   \tab  Grade 12; 1 year of college; 2 years of college;  \cr
      \tab     \tab   \tab  4 years of college; 5+ years of college. \cr
    [,8] \tab EMPSTAT   \tab employment status \tab 4 levels: N/A; Employed; Unemployed; Not in labor force. \cr
    [,9] \tab DISABWRK \tab work disability status \tab 3 levels: N/A; No disability that affects work; \cr
     \tab  \tab  \tab  Disability causes difficulty working. \cr
    [,10] \tab VESTAT \tab veteran status \tab 3 levels: N/A; Not a veteran; Veteran. \cr
}
}

